Made Highly Experimental Version of my Guis. Less code (at all) and less Textures to load but its very slow about (40 Times (or bigger)) but it makes a lot of thing possible :D
Adjusted my Mods (yeah need to adjust the whole mod to it) to it.
Nei Buttons no longer Possible in my Guis Maybe with a fiction class but
i want less classes and nope i do not do that...
And that is it.
Maybe another features:
Guis now Autodraw. It is hardCoded what stays where.
Ctrl/Strg(German Name) on a slot that is empty and it shows a little
info (and yeah it is only possible because of that new gui way at least
in a less pain way). Tanks automaticly show what they have (in guis) and 
they are autodrawen too. No more code to add.
Emmm yeah thats it...
Removed a lot of textures. Removed mostly everything what has todo with
hoppers. (they come back. But not like XYCraft xD)