Another Prerelease Commit:
-Added/Fixed: Metal Mixing Recipes in the Pressure Furnace are
Oredictionaried now. (Some).
-Added some usefull functions to the Redstone Utils.
-Changed: Redstone Updates (receiveing mode) happends on blockUpdates
instead of Ticks, works fine with Vanilla Redstone.
-Added: Some Functions to my GuiCore class so i can add Slider Buttons.
-Added: Scrolling with mouseweel in slider Buttons.
-Fixed: Some display bugs.
-Sorted: Recipes so you can see easier where everything is..
-Finished TinyBarrel.
-Changed: Upgrade/Transforming of TinyChests. So Nei & Craftingguide is
not so overloaded.
-Added: BCBatteries... hmm should call them MJ Batteries in code, they
store MJ and also have the Ability to share Energy thanks to my
ItemEnergyNet which makes a player partly to an energyNet.
-Added: Hemp Items.
-Added: Charging Bench & Battery Station (sadly they do not look as cool
as before in the Gui, sry).
-Added: Textures.
-Added: Batteries raintain their charge when they do get Upgraded.
(After Crafting it charges the battery so do not wonder if the recipes
seem to make a empty battery)
-Fixed: PESU stores now energy again.
-Fixed: Wood Gasser accepts now MFR Wood, also it should accept IC2
Wood. (IF not blame me for that and you get a hotfix)...