Long time no update on TinyModularThings and SpmodAPI and the addon mods, well IC2 Classic came into the way... Now paused a little bit, also did play to much mc, and watched to much animes... Well they are awesome xD.
But todays and last things i changed:
-Packet System is Deprecated. Will changed. If someone uses my API
please look to the changes that will come sooner or later.

-Changed Structure Storage. Instead of Saving Positions it saves now the
box which the Structure want to know if something happens. (Note Player
and Block Boxes are split. So player Interactions will handled differend
then BlockUpdates, makes it easy to say only this area is accessable
through the player, and the structure is completely differend).

-Removed Waila plugin. Its making to much trouble and creating lag, also
i can not compile the code because dev version not working and also
1.6.4. I do not know how much people uses it.

-Fixed Basic AnimalChunkloader and AnimalGas. Working now fine. If gas
still get stuck just a blockupdate and it works fine.

-Added Ventil. Works fine and smooth and also do not tick that much.

-Added: MobMachine Spawner Mode. Activates Mobspawner and can be
compared with soulshards but the it follows the spawnerrules (at most
parts) and also is relying on events also magnum torch dissable it.

-Fixed: Random Trade. If trades were to much they could have sync issus
and completly differend trades you get from that what you want. Simple
solution (took longer then expected), when the server loads the player
it sends a/multible packets with all the aviable trades.

-Fixed: Interfaces. They work now how they should be so no more: "Why
don't you nodest that you are in a structure".

-Fixed: Charging & Discharging bench spaming Small Packets...

-Fixed: TinyBarrel taking infinite items when it is close to full and
less then 1 stack is missing.

-Planed Fix: TinyTanks will be completly overriden because the render
causes infinite amount of lag. Thanks to @Soaryn that he gave me the
idea how to solve it. (Tile&ISimpleBlockRenderer)

-I think thats it.